refactor(platform-navigation): group platform families
