+++ src/app/(tabs)/(messages)/_layout.tsx
@@ -1,6 +1,6 @@
 import { Stack } from 'expo-router';
 
-import { messagesHeaderShown } from '@/frontend/components/navigation/messagesStackPlatform';
+import { messagesHeaderShown } from '@/frontend/components/navigation/messagesStackPlatform/messagesStackPlatform';
 import { useThemeColor } from '@/frontend/hooks/useThemeColor';
 
 export default function MessagesStackLayout() {
