+++ src/frontend/components/navigation/components/PaintingZoomTransition/PaintingZoomTransition.types.ts
