+++ src/frontend/components/navigation/index.ts
@@ -6,4 +6,7 @@ export {
 export { ContextMenuLink } from './components/ContextMenuLink';
 export type { ContextMenuLinkItem, ContextMenuLinkProps } from './components/ContextMenuLink';
 export { NavigationThemeProvider } from './components/NavigationThemeProvider';
-export { PaintingZoomLink, PaintingZoomTarget } from './components/PaintingZoomTransition';
+export {
+  PaintingZoomLink,
+  PaintingZoomTarget,
+} from './components/PaintingZoomTransition/PaintingZoomTransition';
