+++ src/frontend/components/navigation/messagesStackPlatform/messagesStackPlatform.android.ts
