+++ src/frontend/components/navigation/messagesStackPlatform/messagesStackPlatform.ts
