+++ src/frontend/components/navigation/rootStackPlatform/rootStackPlatform.android.ts
