+++ src/frontend/components/navigation/rootStackPlatform/rootStackPlatform.ios.ts
