+++ src/frontend/components/navigation/tabBarPlatform/tabBarPlatform.android.ts
@@ -1,10 +1,10 @@
 // Android's native tab bar accepts image sources, so it uses the Material Symbols PNGs baked by
 // the app-icons generator. iOS selects native SF Symbols in its platform module.
-const homeIcon = require('../../../../packages/app-icons/src/tab-icons/home.png');
-const assistantsIcon = require('../../../../packages/app-icons/src/tab-icons/assistants.png');
-const messagesIcon = require('../../../../packages/app-icons/src/tab-icons/messages.png');
-const settingsIcon = require('../../../../packages/app-icons/src/tab-icons/settings.png');
-const searchIcon = require('../../../../packages/app-icons/src/tab-icons/search.png');
+const homeIcon = require('../../../../../packages/app-icons/src/tab-icons/home.png');
+const assistantsIcon = require('../../../../../packages/app-icons/src/tab-icons/assistants.png');
+const messagesIcon = require('../../../../../packages/app-icons/src/tab-icons/messages.png');
+const settingsIcon = require('../../../../../packages/app-icons/src/tab-icons/settings.png');
+const searchIcon = require('../../../../../packages/app-icons/src/tab-icons/search.png');
 
 export const tabSceneStyle = { height: '100%' } as const;
