+++ src/frontend/components/navigation/tabBarPlatform/tabBarPlatform.ios.ts
