--- src/frontend/features/search/GlobalSearchChrome.tsx
@@ -1,2 +1,0 @@
-export { GlobalSearchChrome } from './GlobalSearchChrome.android';
-export type { GlobalSearchChromeProps } from './globalSearchChrome.types';
