+++ src/frontend/features/search/GlobalSearchChrome/GlobalSearchChrome.tsx
@@ -1,0 +1,2 @@
+export { GlobalSearchChrome } from './GlobalSearchChrome.android';
+export type { GlobalSearchChromeProps } from './GlobalSearchChrome.types';
