+++ src/frontend/features/search/GlobalSearchChrome/GlobalSearchChrome.types.ts
