+++ src/frontend/utils/nativeNavigationMetrics/nativeNavigationMetrics.android.ts
