+++ src/frontend/utils/nativeNavigationMetrics/nativeNavigationMetrics.ios.ts
