+++ src/frontend/utils/nativeNavigationMetrics/nativeNavigationMetrics.ts
